+++ src/main/kotlin/com/github/fantom/codeowners/file/type/kind/BitbucketFileType.kt
@@ -11,7 +11,8 @@ class BitbucketFileType : CodeownersFileType(BitbucketLanguage.INSTANCE) {
         private val subdirectories = setOf(".bitbucket")
     }
 
-    private fun isInSubdir(codeownersFile: VirtualFile) = codeownersFile.parent.name in subdirectories
+    // TODO need to investigate why parent may be null
+    private fun isInSubdir(codeownersFile: VirtualFile) = codeownersFile.parent?.name in subdirectories
 
     override fun getRoot(vcsRoot: VcsRoot, codeownersFile: VirtualFile): VirtualFile? {
         return super.getRoot(vcsRoot, codeownersFile) ?: codeownersFile.parent?.let { codeownersParentDir ->